Use floats instead of double for variant k-mer rates to avoid precision errors in computations (we observed non nul variance when with count=1 which lead to erroneous detected variants).

+++ src/read_analyzer.cpp
@@ -115,7 +115,7 @@ ReadAnalyzer::ReadAnalyzer(double alpha, double threshold, bool weak_mode):
 bool ReadAnalyzer::_firstPassAnalyzis(const ReadAnalyzer::VariantKmerRatesIteratorWrapper &vr, ReadAnalyzer::VariantStatistics &stats) {
   assert(!_completed);
   // update mean and variance using the Welford's algorithm
-  double delta = vr.rate - stats.mean;
+  float delta = vr.rate - stats.mean;
   ++stats.count;
   stats.mean += delta / stats.count;
   stats.variance += (vr.rate - stats.mean) * delta;
